+++ lib/telegram/bot/botan.rb
@@ -3,24 +3,37 @@ module Telegram
     class Botan
       TRACK_URI = 'https://api.botan.io/track'.freeze
 
+      autoload :ClientHelpers, 'telegram/bot/botan/client_helpers'
+      autoload :ControllerHelpers, 'telegram/bot/botan/controller_helpers'
       class Error < Bot::Error; end
 
+      extend Initializers
+      prepend Async
       include DebugClient
+
+      class << self
+        def by_id(id)
+          Telegram.botans[id]
+        end
+
+        def prepare_async_args(method, uri, query = {}, body = nil)
+          [method.to_s, uri.to_s, Async.prepare_hash(query), body]
+        end
+      end
 
       attr_reader :client, :token
 
-      def initialize(token)
+      def initialize(token = nil, **options)
         @client = HTTPClient.new
-        @token = token
+        @token = token || options[:token]
       end
 
       def track(event, uid, payload = {})
-        res = http_request(
-          :post,
-          TRACK_URI,
-          {token: token, name: event, uid: uid},
-          payload.to_json,
-        )
+        request(:post, TRACK_URI, {name: event, uid: uid}, payload.to_json)
+      end
+
+      def request(method, uri, query = {}, body = nil)
+        res = http_request(method, uri, query.merge(token: token), body)
         status = res.status
         return JSON.parse(res.body) if 300 > status
         result = JSON.parse(res.body) rescue nil # rubocop:disable RescueModifier
@@ -30,6 +43,10 @@ module Telegram
 
       def http_request(method, uri, query, body)
         client.request(method, uri, query, body)
+      end
+
+      def inspect
+        "#<#{self.class.name}##{object_id}(#{@id})>"
       end
     end
   end