+++ lib/telegram/bot/botan/client_helpers.rb
@@ -1,0 +1,15 @@
+module Telegram
+  module Bot
+    class Botan
+      # Helpers for botan.io metrics.
+      module ClientHelpers
+        attr_reader :botan
+
+        def initialize(*, botan: nil, **)
+          super
+          @botan = Botan.wrap(botan, id: id) if botan
+        end
+      end
+    end
+  end
+end